+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 
 All notable changes to FundOS are documented here.
 
+## [1.0.3] — 2026-08-22
+
+### Fixed
+
+- Update OFX parser to stop treating closing tags in transactions as malformed
+- Update OFX parser to handle self closing tags more gracefully
+- Update import process to allow user selection of name or memo field for imported memo
+
 ## [1.0.2] — 2026-07-06
 
 ### Fixed